refactor: iputil package を utils から分離
