+++ utils/iputil/private_ip.go
@@ -1,5 +1,4 @@
-// revive:disable-next-line FIXME: https://github.com/traPtitech/traQ/issues/2717
-package utils
+package iputil
 
 import "net"
